extracting not found into a separate component
